Start refactor CPUAndThreadUtils make it available to EstimationReader
